refactor: move the components folder to the src folder
