+++ src/app/(medium-layout)/admin/(add-edit-project)/AddEditProjectTemplate.tsx
@@ -1,7 +1,7 @@
 'use client'
-import Button from '@/app/components/Button'
-import Input from '@/app/components/Input'
-import RadioInputs from '@/app/components/RadioInputs'
+import Button from '@/components/Button'
+import Input from '@/components/Input'
+import RadioInputs from '@/components/RadioInputs'
 import useSubmitForm from '@/utils/useSubmitForm'
 import 'bootstrap-icons/font/bootstrap-icons.css'
 import { useRouter } from 'next/navigation'
