+++ src/app/(medium-layout)/page.tsx
@@ -1,4 +1,4 @@
-import ProjectsPageTemplate from '@/app/components/ProjectsPageTemplate'
+import ProjectsPageTemplate from '@/components/ProjectsPageTemplate'
 
 export default function ProjectsPage() {
     return <ProjectsPageTemplate getProjectsEndpoint="/public/projects/all" />
